Add length validation of post content
This commit adds validation of length of $content property of an instance of
Comment and Article. Since the property is inherited from their common
subclass: AbstractMappedPost, the validation is added to the subclass, even
though criteria of validity are separate. This is handled by adding two
separate assertions in separate validation groups.