Assign some validation constraints to "UnregisteredAuthor" validation group
This commit assigns validation of properties authorName and authorEmail
of Comment class to "UnregisteredAuthor" validation group. Their validation
will be performed only when the author of a comment is not registered, because
only in that case the properties can (or must, in case of authorName) be set
to values other than null.